chore: scc folder renamed
