+++ css/SignUpTwoScreenStyles.tsx
